requirements.txt
python-telegram-bot==22.0
httpx==0.28.1
beautifulsoup4==4.12.3
APScheduler==3.10.4
openai==1.68.2
python-dotenv==1.0.1